ping_individual_4_ports: Skip master packets counter check on SPU3
The SPU3 does not have a dedicated management Ethernet interface. As a
result, one of the switch ports has to be used. The ssh sessions to
get the statistics counters get counted in the statistics counters for
the master interface, so we cannot test if this counter is as
expected.