bin/wemo-switch: Python script to control Belikin wemo switches
The default wemo tools, as part of Ouimeaux, have problems with the
switch 'disappearing'. This appears to be a uPnP issue, maybe the WiFi
AP not forwarding multicast correctly. The uPnP discovery can however
be bypassed and you can manipulate the device directly, if you know
its IP address/hostname. This script does this, making controlling the
switch much more reliable.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>