ping_individual_4_ports_test: Test big packets pings
There has been a number of boards with MTU problems, not allowing a
fill size frame, plus DSA tag to be received. Ping using a big packet,
which should get fragmented into a full MTU packet, and a shorter packet.
If this does not work, there is likely to be an MTU issue on the target.